Adds frequency–inverse document frequency calculation Changes Longs to be Integer for faster calculation Replaces UniqWordsExtractor dictionary to be a linkedHashSet so that the order is respected

+++ src/main/java/Model/IndexRecord.java
@@ -12,7 +12,7 @@ import java.util.UUID;
 public class IndexRecord {
     // TODO: Change vector into something that are aware of which extractor is creating the part
     @NonNull
-    private List<Double> vector;
+    private List<Integer> vector;
     @NonNull
     private String name;
     @NonNull
